refactor: extract mixins for base manager
